build: Correct multiple signed/unsigned comparison errors
Resolves numerous -Wsign-compare build errors flagged by stricter
compiler settings.

The most critical change involves refactoring read() and write()
calls to safely handle their signed ssize_t return values. The new
pattern prevents latent bugs by checking for system errors (-1)
before checking for partial I/O.

This commit also corrects other sign-compare errors by applying
explicit casts where appropriate.

+++ ici/sdr/sdrhash.c
@@ -84,7 +84,7 @@ Object	Sdr_hash_create(const char *file, int line, Sdr sdrv, int keyLength,
 		rowCount = hashDimensions[i];
 		i++;
 		if (rowCount <= rawRowCount
-		&& i < (sizeof hashDimensions / sizeof(int)))
+		&& (size_t)i < (sizeof hashDimensions / sizeof(int)))
 		{
 			continue;
 		}
@@ -533,7 +533,7 @@ static void	deleteHashEntry(Sdr sdrv, Object eltData, void *arg)
 {
 	/* Parameter intentionally unused. */
 	(void)arg;
-	
+
 	sdr_free(sdrv, eltData);
 }
